requirements.txt
aiogram==3.15.0
asyncpg==0.30.0
openai==1.82.0
structlog==24.4.0
aiohttp==3.11.16
python-dotenv==1.1.0
uvicorn==0.34.2
fastapi==0.115.12
gunicorn==23.0.0
